backend/requirements.txt
anthropic==0.28.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
boto3==1.34.0
python-multipart==0.0.6
